fix minor compiler warnings
svn path=/trunk/; revision=28207

+++ applications/smurf/libsmf/smf_sortd.c
@@ -79,6 +79,8 @@
 #include "sae_par.h"
 #include "ast.h"
 
+#include "smf.h"
+
 int *smf_sortd( size_t nel, double *array, int *sorted, int *status ){
 
 /* Local Variables */
@@ -89,7 +91,7 @@ int *smf_sortd( size_t nel, double *array, int *sorted, int *status ){
    int *r = NULL;
    int *result = NULL;
    int done;
-   int i;
+   size_t i;
    int temp;
 
 /* Check the inherited status. */
